LayoutTests/storage/indexeddb/index-get-key-argument-required-expected.txt

Test IndexedDB index .get() required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo', { keyPath: 'id', autoIncrement: true });
index = objectStore.createIndex('first', 'first');
PASS index.get(); threw exception TypeError: Not enough arguments.
PASS index.getKey(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
